final commit before submitting bug. moved some classes to consolidate folders
